Fix function parameter value overflow that caused problems when copying data from a GDAL dataset. Problem first appeared in ST_Resize(). Ticket #2188
git-svn-id: http://svn.osgeo.org/postgis/trunk@11057 b70326c6-7e19-0410-871a-916f4a2858ee

+++ raster/rt_core/rt_api.h
@@ -638,7 +638,7 @@ rt_errorstate rt_band_get_nodata(rt_band band, double *nodata);
 rt_errorstate rt_band_set_pixel_line(
 	rt_band band,
 	int x, int y,
-	void *vals, uint16_t len
+	void *vals, uint32_t len
 );
 
 /**
